refactor: move storageAdapter to dedicated file and update imports
